backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pandas==2.1.3
numpy==1.26.2
networkx==3.2.1
python-multipart==0.0.6
reportlab==4.0.7
pyvis==0.3.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
scikit-learn==1.3.2
scipy==1.11.4
matplotlib==3.8.2
bcrypt==4.1.2
cryptography==41.0.7
PyJWT==2.8.0
httpx==0.25.2
redis==5.0.1
prometheus-client==0.19.0
python-decouple==3.8
email-validator==2.1.0
slowapi==0.1.9
certifi==2023.11.17